Fix reload memory leak
Each channel structure includes a list of workers to notify
(workers_with_subscriptions) when new messages arrive for the channel,
the exit of a worker does not remove it's entry from the subscriptions list.
This means that a receiving worker copies a new message to a worker which
no longer exists and the message is never processed.

If a new worker is occupying the slot where an old one was it will simply
discard messages not intended for it. However, when no worker is on a slot
for which messages arrive they simply accumulate and leak all our memory away.